+++ cdk.json
@@ -1,5 +1,5 @@
 {
-  "app": "npx ts-node bin/playground-infra.ts",
+  "app": "npx ts-node bin/boomtap-infra.ts",
   "context": {
     "@aws-cdk/core:enableStackNameDuplicates": "true",
     "aws-cdk:enableDiffNoFail": "true",
