+++ test/boomtap-infra.test.ts
@@ -4,12 +4,12 @@ import {
   MatchStyle,
 } from "@aws-cdk/assert";
 import * as cdk from "@aws-cdk/core";
-import * as PlaygroundInfra from "../lib/playground-infra-stack";
+import * as BoomtapInfra from "../lib/boomtap-infra-stack";
 
 test("Empty Stack", () => {
   const app = new cdk.App();
   // WHEN
-  const stack = new PlaygroundInfra.FrontEndStack(app, "MyTestStack");
+  const stack = new BoomtapInfra.FrontEndStack(app, "MyTestStack");
   // THEN
   expectCDK(stack).to(
     matchTemplate(
